+++ pkg/apis/infrastructure/v1alpha2/g8s_control_plane_types.go
@@ -59,6 +59,7 @@ type G8sControlPlaneSpec struct {
 
 // G8sControlPlaneStatus defines the observed state of G8sControlPlane.
 type G8sControlPlaneStatus struct {
+	// +kubebuilder:validation:Enum=1;3
 	// Total number of non-terminated machines targeted by this control plane
 	// (their labels match the selector).
 	// +optional